Move CLI timeSync helper into cli/lib/timing.js
Relocated timeSync from src/shared/number-utils.js to src/cli/lib/timing.js because it logs CLI-specific verbose output. Updated the manual generation commands to import the helper and progress bar utilities from cli/lib so shared utilities stay domain-neutral.
